Fix the overall solve rate display not correctly being calculated.

+++ Assets/Scripts/UI/TwitchLeaderboard.cs
@@ -59,7 +59,7 @@ public class TwitchLeaderboard : MonoBehaviour
         }
         else
         {
-            totalSolveRate = solveCount / strikeCount;
+            totalSolveRate = ((float)solveCount) / strikeCount;
         }
 
         totalBombCountText.text = bombCount.ToString();
